requirements.txt
annotated-types==0.7.0
anyio==4.12.1
arrow==1.4.0
asgiref==3.11.0
binaryornot==0.4.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
distro==1.9.0
Django==6.0.1
django-admin-interface==0.32.0
django-browser-reload==1.21.0
django-ckeditor==6.7.3
django-colorfield==0.14.0
django-js-asset==3.1.2
django-tailwind==4.4.2
gunicorn==25.0.3
h11==0.16.0
honcho==2.0.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openai==2.20.0
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
text-unidecode==1.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
